[pinpoint] Define some Quests.
Placeholders for things that will run on a Change. Quests will return two types of results: Values and parameters that are passed to the next Quest in the execution chain. (With Dungeon Master, the second thing will be part of DM's dependency system, so we won't need to worry about it.)

BUG=catapult:#2891

Review-Url: https://codereview.chromium.org/2429013003