Helper function for get interactive timestamps
To compute Estimated Input Latency, we need to identify task windows after a
page becomes interactive and until the user navigates away from the page. When a
page becomes interactive is defined by the TTI metric. This CL adds a function
to extract all interactive timestamps of a single renderer.

BUG=chromium:631206

Review-Url: https://chromiumcodereview.appspot.com/2428953002