Refactor histogram_set_table.html.
Currently, histogram_set_table.html is a huge file containing 3 different
dom-modules.

This CL splits out the table-cell and table-name-cell to their own files,
and defines an ES6 class HistogramSetTableRow to contain all of the
row-specific logic, and ES6 Maps instead of dictionaries.

I limited myself to only fixing one small bug :-)
Filtering by the search box is currently incorrectly done after the rows are built,
so it only works when |name| is the first grouping.
This patch moves it to before the histograms are grouped so that search always
works regardless of grouping keys.

Review-Url: https://codereview.chromium.org/2472803004