Override overview line charts dataRange.
Currently, all of the overview line charts in histogram-set-table
compute their own dataRanges.
This makes it difficult to scan across a row, because you have to stop and read
each y-axis.

This CL uses the new chart overrideDataRange to synchronize the y-axes of all of
the overview line charts in each row.

https://screenshot.googleplex.com/U967DhfyMcu.png

Follow-up CLs will
 * fix the x-axes to be like NameColumnChart,
 * add hoverable dots, and
 * fix the sizing issue.

Review-Url: https://codereview.chromium.org/2491213004