[dashboard] Fix use of () => {}
While () => [] returns a new array very unexpectedly
() => {} doesn't return a new object and instead returns
undefined. Since the arrow function also has a statement variation which
looks like: () => { foo(); return bar(); } the {} in () => {} is
interpreted as an empty statement block without a return so we get
undefined. This is almost certainly not what was expected.

BUG=catapult:#3039

Review-Url: https://codereview.chromium.org/2531293002