Plumb label, benchmarkStartMs thru ChartJsonConverter.
A recent CL converted chart-json to Histograms so that TBMv1 and legacy telemetry
benchmarks could produce results2 so that we can delete results1.
https://codereview.chromium.org/2474573002

That CL forgot to plumb --results-label and benchmarkStartMs, so results were not
splayed into columns properly.

This CL plumbs label and benchmarkStartMs through ChartJsonConverter so that TBMv1
and legacy benchmark results are splayed into columns in results2.

https://x20web.corp.google.com/~benjhayden/2531423002.html

BUG=catapult:#3036

Review-Url: https://codereview.chromium.org/2531423002