[Telemetry] Fix JavaScript interpolation in telemetry actions
Telemetry actions expose API's for clients to interact with elements on
a page (e.g. scroll or tap). Some of these take parameters (e.g. a CSS
selector) which were naively interpolated into JavaScript snippets, but
this is brittle when special characters are involved.

In this CL we migrate all of these to use the js_template helper, which
takes care of safely inserting Python values into the JavaScript
snippets.

BUG=catapult:#3028

Review-Url: https://codereview.chromium.org/2559503002