Add BuildbotInfo Diagnostic.
Currently, TelemetryInfo contains information from telemetry and the trace. We
have plans to plumb more kinds of information from more sources through the
system. Instead of expanding TelemetryInfo, we're adding a few new different
kinds of Diagnostics.

BuildbotInfo will be added as a shared Diagnostic to a HistogramSet by a process
outside of telemetry. A Python module is provided to assist that process.

Review-Url: https://codereview.chromium.org/2566493002