Changing group reports to use an id for multiple keys.
When a user selects alerts (from the alerts/ page), their keys are sent (via post) to /group_report which replies with an id (a hash of the keys, which maps to the keys in the datastore).
Then, we redirect to /group_report?id= which (in the ready: function) posts to /group_report_data with the id and receives the applicable keys (and all the other information from before).

Also handles ensuring the correct alerts are selected. Adds corresponding tests.
Note, the original ?keys=, ?bug_id=, and ?rev= work as before.

BUG=catapult:#2876

Review-Url: https://codereview.chromium.org/2537053003