[devil] Update location of command line file on devices
Chrome now expects the command line file to be written at
/data/local/tmp.

API change: clients no longer need to specify the full path to the
command line file, just the file name is enough. To ease the transition,
if a full path is given the directory is stripped off and suitable
warnings are emitted.

This change also fixes an issue where Telemetry wouldn't work due to
legacy command line files placed on the wrong location.

BUG=catapult:#3112

Review-Url: https://codereview.chromium.org/2625743002