Add MergedTelemetryInfo.equals().
When merging TelemetryInfos into MergedTelemetryInfos, it is possible for there
to be multiple instances of MergedTelemetryInfos with the same values.
It will be more efficient to re-serialize such diagnostics, and use them to
merge histogram relationships, if these equivalent diagnostics are deduplicated.
In order to deduplicate diagnostics, we will need some way of determining if
they are equivalent.
Forthcoming MergedBuildbotInfo and MergedDeviceInfo will be very similar to this.

For more context, this CL is split out from
https://codereview.chromium.org/2683463002

BUG=catapult:#2842

Review-Url: https://codereview.chromium.org/2692263002