[Dashboard] Add 'test_path_dict' request type to /list_tests
This CL adds a minimal test_path_dict request type to list_tests as part of the
program to implement https://codereview.chromium.org/2647343009.

We start by considering the case where we are returning selected tests and the
tests are specified concretely, or exhaustively ('all').

Followup CLs will broaden this API surface as follows:
1. Support unselected tests
2. Support computation of core tests.

We will also be able to begin using this functionality in /report,
/group_report, and /graph_data as it lands.

BUG=catapult:#2887

Review-Url: https://codereview.chromium.org/2698583002