[Dashboard] Only validate row_id for first row in add_point
Previously, we were validating the row_id for every single row to be added
during an add_point request. The validation process requires us to perform
gets, which meant that there was at least one get for every row. However, all
points added during an upload can be assumed to share the same row_id, so we
really only have to perform that validation once.

BUG=catapult:#3108

Review-Url: https://codereview.chromium.org/2686963003