If a user supplies 1 or no revisions (and no milestone param) still do something reasonable.
If no rev/milestones supplied, display most recent milestone - most recent revision.
If 1 rev supplied, display that revision - end of milestone that revision is in.
If 2 rev supplied, display those revisions.
If milestone supplied, display that milestone.

Demo (logged in):
https://dev-jessimb-767bc124-dot-chromeperf.appspot.com/speed_releasing/MemoryTest7?revA=1479546161&revB=1485025126
is the same as
https://dev-jessimb-767bc124-dot-chromeperf.appspot.com/speed_releasing/MemoryTest7?revB=1479546161
is the same as
https://dev-jessimb-767bc124-dot-chromeperf.appspot.com/speed_releasing/MemoryTest7?m=56

And also,
https://dev-jessimb-767bc124-dot-chromeperf.appspot.com/speed_releasing/MemoryTest7?m=57
is the same as
https://dev-jessimb-767bc124-dot-chromeperf.appspot.com/speed_releasing/MemoryTest7

BUG=catapult:#3141

Review-Url: https://codereview.chromium.org/2689023002