[Dashboard] Implement unselected case for 'core' selection
Following on from https://codereview.chromium.org/2712813002, this CL is the
last piece of the puzzle for the test_path_dict mode of list_tests: it should
now be able to service all existing use cases.

Further CLs will begin to use this endpoint:

  * In /group_report
  * In /report
  * In chart-container.

Once that work is done, it should be possible to eliminate test-picker's
request for the subtest dict.

BUG=catapult:#3225

Review-Url: https://codereview.chromium.org/2715433005