Fix histogram-set-table expanding name column.
Currently, results.html constrains the name column by default, but fails to show
the dots button to expand the column.
This is a simple timing bug: results2_template.html configures the
histogram-set-table, then displays it, so when the name-cells call
getComputedStyle() to find their width, it returns 0 so they think they are not
too long, so they don't fire name-cell-overflow, so the histogram-set-table does
not show the dots button.
This bug does not show up in the tests because the tests don't hide the table,
then configure it, then display it like results2_template.html does.

This CL adds a polymer-style callback "displayed()" to histogram-set-table and
makes results2_template.html call it after displaying the table so that it can
iterate over the rows and see if any of the name cells overflow, and show the
dots button if they do.

Small parts of this CL will be replaced by #3289 but most of this CL will
remain after that work is complete, and this bug fix is necessary while that work
is ongoing.

BUG=catapult:#2946,catapult:#3289

Review-Url: https://codereview.chromium.org/2737013002