[Devil] Run commands in flag_changer forwarder and perf_control as shell commands.
They are being passed as strings, so they should set shell=True. Currnetly not having
this set is causing a lot of log spam in telemetry runs.

WARNING:root:The command to run should preferably be passed as a sequence...
This warning message shows up 1600+ times in system_health runs.

BUG=chromium:666040

Review-Url: https://codereview.chromium.org/2746153009