Rebin single-bin Histograms.
Currently, legacy benchmarks produce chart-json, which is converted to
single-bin Histograms, which are displayed as box plots, which confuse some
users and are not currently synchronized across rows.

This CL rebins all histograms in a row if they are all single-bin Histograms.
The new bin boundaries are a linear interpolation between the power of 10 below
the smallest sample in the row and the power of 10 above the largest sample in
the row.
All of the sampleValues from the single-bin Histogram are added to the rebinned
Histogram, but then the statistics are replaced with the single-bin Histogram's
statistics in case the single-bin Histogram dropped samples.
This CL ensures that single-bin Histograms have at least 400 samples (40 bins in
the rebinned Histogram * 10 average samples per bin)

Without rebinning: http://www/~benjhayden/2531423002.html
With rebinning: http://www/~benjhayden/2745113004.html
(defaultMaxNumSampleValues was 10 when those files were produced. After this CL,
legacy benchmarks will serialize up 400 sampleValues per Histogram, which will
fill out the bins much better than that file can demonstrate.)

BUG=catapult:#3288

Review-Url: https://codereview.chromium.org/2745113004