[pinpoint] Add Change.Midpoint() and Dep.Midpoint()
These classmethods are used for bisection.

Additionally, add __str__ and __eq__ methods for debugging and testing.

Previously, Change internally kept track of base_commit and additional deps
separately. For a lot of the logic, and especially the Midpoint logic, it's
cleaner to keep track of them together. So now we have Change._all_deps.

BUG=catapult:#3332

Review-Url: https://codereview.chromium.org/2727383005