ViewState is a baseclass for MVC Models.
Subclasses declaratively define properties.
When properties are set, an update event is dispatched so that multiple Views
and Controllers can subscribe to update events on the same models.

Subclasses can contain instances of ViewState subclasses in Arrays, Sets, or
Maps in order to build tree or even general graph structures.
ViewStates are serializable.

When d8 is rolled on all platforms, a followup will make update() async and await
on all listeners so they can be async as well.

Followups will convert Range, RunningStatistics, Histogram, and the Diagnostics to
subclass Serializable.

BUG=catapult:#3289

Review-Url: https://codereview.chromium.org/2761123003