Add getter for thread type
For the new cpu time metric, we would like to group threads with names like
"WorkerPool1/23" and "WorkerPool1/34" as the same "type". This CL adds a getter
for the type of a thread. The type is purely a function of the thread name, and
there is no way to set it manually.

R=benjhayden@chromium.org,tdresser@chromium.org

BUG=catapult:#3325

Review-Url: https://codereview.chromium.org/2800933002