Move the feedback form from group-report-page to report-page.
A feedback form is not necessary for group-report-page because that page is mostly
used by sheriffs, who already receive a form.

Some questions on the form are also changed. The new form is at
https://goo.gl/forms/BJM8ZKW3m8ZC7tYZ2
The new form asks that bugs be filed directly to github.

Demo: https://dev-benjhayden-b0384fe6-dot-chromeperf.appspot.com/report

http://i.imgur.com/qgyln4R.png
Review-Url: https://codereview.chromium.org/2917213002