Fix histogram-span for Histograms with a single sample with diagnostics.
Currently, if a histogram contains a single sample with sample diagnostics,
histogram-span correctly decides not to display the bar chart, but fails to
display the sample diagnostic.
This CL refactors the plumbing for brushedBinRange and mergeSampleDiagnostics to
prevent that kind of bug.

In fixing that bug, I noticed that (non-RelatedHistogram) Breakdowns logged errors
on the console when mousing over its stacked column chart, because
keySpan.parentNode was sometimes undefined. Breakdown shouldn't try to use
parentNode.nextSibling, it should be able to rely on displayElement being an
Element.

Review-Url: https://codereview.chromium.org/2931973002