Adds option to not have SetLogLevel add a log handler.
I want to add a custom log formatter with color to Chromium's
test_runner. To do so, I will need to add my own log handler and
do not want the one added here. It seems strange that SetLogLevel
adds a handler in the first place. So adding option to disable
this behavior.

Review-Url: https://codereview.chromium.org/2917343005