[Tracing] Add HistogramSet.GetSharedDiagnosticsOfType
This CL adds a helper for an increasingly common pattern in Python code for
dealing with shared diagnostics: querying for all the shared diagnostics of a
given type. Two examples of where this operation is necessary are the
dashboard's deduplication logic for sparse diagnostics, which must first match
diagnostics based on type, and unit tests for code that adds shared diagnostics
to HistogramSets, where it is useful to be able to find matching diagnostics
without having to traverse histograms.

Review-Url: https://codereview.chromium.org/2933083003