Add a slider to histogram-set-controls to control alpha.
Currently, the meaning of the emoji can sometimes be unclear because the concept
of the alpha threshold is entirely hidden.
If users had more visibility into and control over the alpha threshold, then it
would be easier for users to interpret the emoji.
This CL adds a slider to allow users to select custom values of alpha between
0.001 and 0.1.

https://benjhayden.users.x20web.corp.google.com/www/2933183002.html?r=blink_perf.css%0A2017-06-08%2016%3A50%3A53&s=%CE%94avg&g=name&p=0.02

http://i.imgur.com/g0mKjUH.png
Review-Url: https://codereview.chromium.org/2933183002