Manage focus in overlays.
Currently, whenever any elements other than the overlay <div> itself receives
focus, the overlay immediately steals it back.
https://github.com/catapult-project/catapult/commit/22cfc4803f75e4bf370c50dedc3a143934df227d

This is ok for radio buttons and checkboxes, but catastrophic for select boxes,
which can only be interacted with while focused.

This CL makes overlay only steal focus back when the user tries to focus on
something outside of the overlay. This allows select boxes inside of the overlay
to maintain focus so that users can interact with them.

I can't figure out why this ever worked. Maybe Chrome behavior changed?

BUG=chromium:733991

Review-Url: https://codereview.chromium.org/2946513002