symbolize_trace: Always symbolize all symbols by default.
When a user is symbolizing a trace, assume that all symbols are present. If all
symbols are not present [e.g. because the trace is coming from another user's
machine], assume that the caller of the script is sufficiently familiar with the
process that they will pass the flag --only-symbolize-chrome-symbols.

BUG=chromium:737658

Review-Url: https://codereview.chromium.org/2960243002