Fix CSVBuilder.
CSVBuilder currently assumes that all Histograms have a TelemetryInfo, not a
MergedTelemetryInfo, so it's broken for merged Histograms.
Also, TelemetryInfo is about to be refactored.

This CL makes CSVBuilder use HistogramGroupings instead of accessing diagnostics
directly.
When TelemetryInfo is refactored, the HistogramGroupings will need to be
updated, but not CSVBuilder.

BUG=catapult:#3507,catapult:#3582

Review-Url: https://codereview.chromium.org/2975863002