Ungroup unrelated memory alerts.
Currently, all alerts for a given benchmark are grouped together if their revision
ranges overlap, even if the measurements are unrelated.
This is particularly troublesome for memory metrics.

This CL hard-codes memory metric relationships in the dashboard frontend in order
to prevent grouping together alerts for unrelated memory metrics.

We have a plan to generalize this to support all TBMv2 metrics' relationships
(without hard-coding relationships in the dashboard frontend) using the
Histogram Pipeline and 'related names' diagnostic.

https://dev-benjhayden-d5329c1a-dot-chromeperf.appspot.com/alerts?sortby=end_revision&sortdirection=down&sheriff=Chromium%20Perf%20Sheriff&triaged=true
https://dev-benjhayden-d5329c1a-dot-chromeperf.appspot.com/alerts?sortby=end_revision&sortdirection=down&sheriff=Desktop%20Memory%20Sheriff
https://dev-benjhayden-d5329c1a-dot-chromeperf.appspot.com/alerts?sortby=end_revision&sortdirection=down&sheriff=Blink%20Memory%20Mobile%20Sheriff

BUG=catapult:#3651

Review-Url: https://codereview.chromium.org/2973883002