[Telemetry] Rename PermanentlyDisableBenchmark to DisableBenchmark.
Benchmarks now have a variable (SUPPORTED_PLATFORMS) that is used to
determine if a benchmark should run on a given platform. DisableBenchmark
will be used in cases where we want to disable the benchmark, but want
it to be clear that its disabled because it is failing not becuase it is
not intended to run on that platform.

TBR=nednguyen@google.com
BUG=chromium:713222

Review-Url: https://codereview.chromium.org/3000413002