webrtc: Fix missing port forwarding on CrOS
There is a local streaming server started in webrtc benchmark for
CrOS browser to connect to, however, the http server is not on the
CrOS device and is not reachable from the device without remote
port forwarding. The patch addresses this issue by creating a
remote port forwarder if the local server and device are not on
the same machine.

BUG=chromium:740900
TEST=run_benchmark --browser=cros-chrome --output-format=chartjson --remote=DUT_IP webrtc

Signed-off-by: Chung-yih Wang <cywang@chromium.org>
Review-Url: https://codereview.chromium.org/2982743002